add query db

+++ entity/categories.go
@@ -1,7 +1,7 @@
 package entity
 
 type Categories struct {
-	CategoryId int32
+	CategoryId int64
 	CategoryName string
 	Description string
 	Picture []byte
